// src/types/colors.ts
export type Colors = {
  primary: string;
  white: string;
  black: string;
  green: {
    50: string;
    100: string;
    200: string;
    300: string;
    400: string;
    500: string;
    600: string;
    700: string;
    800: string;
    900: string;
  };
  grey: {
    50: string;
    100: string;
    200: string;
    300: string;
    400: string;
    500: string;
  };
  info: {
    50: string;
    100: string;
    200: string;
  };
  yellow: {
    200: string;
    400: string;
  };
  orange: {
    50: string;
    100: string;
    200: string;
  };
  red: {
    50: string;
    100: string;
    150: string;
    200: string;
    300: string;
    400: string;
    500: string;
    550: string;
    600: string;
    700: string;
    800: string;
    900: string;
  };
  slate: {
    800: string;
    900: string;
  };
};